refactor: remove core
