+++ test/helpers/check-is-balanced.ts
@@ -1,7 +1,10 @@
 import { updatePendingPayments } from "@app/payments"
 import { updatePendingInvoices, updateOnChainReceipt } from "@app/wallets"
-import { balanceSheetIsBalanced } from "@core/balance-sheet"
 import { baseLogger } from "@services/logger"
+import { getBalance as getBitcoindBalance } from "@services/bitcoind"
+
+import { ledgerAdmin } from "@services/mongodb"
+import { lndsBalances } from "@services/lnd/utils"
 
 import { waitUntilChannelBalanceSyncAll } from "./lightning"
 
@@ -23,4 +26,51 @@ export const checkIsBalanced = async () => {
 
   // TODO: need to go from sats to msats to properly account for every msats spent
   expect(Math.abs(bookingVersusRealWorldAssets)).toBe(0)
+}
+
+const getLedgerAccounts = async () => {
+  const [assets, liabilities, lightning, bitcoin, bankOwnerBalance] = await Promise.all([
+    ledgerAdmin.getAssetsBalance(),
+    ledgerAdmin.getLiabilitiesBalance(),
+    ledgerAdmin.getLndBalance(),
+    ledgerAdmin.getBitcoindBalance(),
+    ledgerAdmin.getBankOwnerBalance(),
+  ])
+
+  return { assets, liabilities, lightning, bitcoin, bankOwnerBalance }
+}
+
+const balanceSheetIsBalanced = async () => {
+  const { assets, liabilities, lightning, bitcoin, bankOwnerBalance } =
+    await getLedgerAccounts()
+  const { total: lnd } = await lndsBalances() // doesnt include escrow amount
+
+  const bitcoind = await getBitcoindBalance()
+
+  const assetsLiabilitiesDifference =
+    assets /* assets is ___ */ + liabilities /* liabilities is ___ */
+
+  const bookingVersusRealWorldAssets =
+    lnd + // physical assets
+    bitcoind + // physical assets
+    (lightning + bitcoin) // value in accounting
+
+  if (!!bookingVersusRealWorldAssets || !!assetsLiabilitiesDifference) {
+    logger.warn(
+      {
+        assetsLiabilitiesDifference,
+        bookingVersusRealWorldAssets,
+        assets,
+        liabilities,
+        bankOwnerBalance,
+        lnd,
+        lightning,
+        bitcoind,
+        bitcoin,
+      },
+      `not balanced`,
+    )
+  }
+
+  return { assetsLiabilitiesDifference, bookingVersusRealWorldAssets }
 }
